tris: Add initial client interoperability tests.
Prepare framework for testing tls-tris as client against other servers.
Currently only boringssl is implemented, but the idea is to add support
for others too (NSS, OpenSSL, picotls, tris, ...).

ecdsa.pem, rsa.pem are copied from tris-localserver. The boringssl image
is reused for the server since the binary was built anyway, maybe this
will be changed if it turns out to be unmaintainable.

Note: the boringssl version is also bumped to something closer to master
to fix a build error and  make the -loop and -www options work.