tris: Advertise PSS in Client Hello for TLS 1.3
Until PSS certificates are properly implemented, let's not advertise
support for that. Since TLS 1.3 however mandates PSS, we have no other
option than advertising this even if we have not added complete support.

Another reason why I apply it to just TLS 1.3 and not 1.2 is because the
latter would require updating the testdata.