Create package crypto/tls/constants
The DC extension has its own package (crypto/tls/delegated_credential).
This is useful because its API is not only invoked by the TLS client and
server, but also by the backend delegator (the endity that signs
delegated credentials for use by the TLS server). Thus, if we wanted to
implement DC in the crypto/tls package, we would need to change its
exported API. It seems cleaner to leave this API alone.

The problem is that the DC extension needs to import some constants from
crypto/tls: in particular the identifiers for signature schemes and the
protocol version. But since crypto/tls imports
crypto/tls/delegated_credential, there is a circular dependency. This PR
removes this circular dependency by defining these constants in a new
package crypto/tls/constants.