travis: speed up test-unit on Linux, rewrite test matrix
Apparently specifying multiple Go versions does not exactly work when
using a matrix with includes, it would add new entries to the matrix
instead of duplicating them. Move to the original implicit matrix and
exclude combinations that do not work (osx with Docker).

While at it, skip pip install docker when Docker is not needed, this
saves about 10 minutes.