interop_test_runner: use Python 3, improve wait logic
'python2' does not exist on macOS, it is called 'python2.7'. Porting to
Python 3 is however more future-proof so do that instead. Rename the
file such that it can be run with pytest.

Do not listen on a fixed host port and use a random port instead.
Otherwise tests could fail if something is listening on those ports.