interop_test_runner: speed up tests by not waiting for three seconds
Do not wait three seconds for every server and instead speed it up by
checking for for availability of services.

Starting a container just to check for ports is a bit silly though, by
the time it is started the target container is probably already. Another
option is to run 'ss -Htln', but exec_run in docker-py 3.7.2 has a bug
that causes ugly file handle leaks.