Makefile: fix vendoring issue for sike and sidh dependencies
Do not vendor github.com/cloudflare/sike and github.com/henrydcase/nobs
in Go 1.12 and before, that does not work when using -mod=vendor.
Instead use the 'internal/' prefix as is done for golang.org/x/ packages
in Go 1.12. For Go 1.13, the vendor directory can be used, so do so.

Fixes #177